refactor: 최대 쿠폰 개수 상수화
